+++ OniAccess/OniAccessStrings.cs
@@ -1317,6 +1317,7 @@ namespace STRINGS {
 			public static LocString TILE_PLURAL = "tiles";
 			public static LocString NO_CHANGE_BOUNDARY = "no change till map boundary";
 			public static LocString HELP_SKIP = "Skip cursor to next change";
+			public static LocString HELP_SKIP_DEFAULT = "Skip by building or element (ignore overlay)";
 		}
 
 		// Colony diagnostics readout and browser